refactor(core): consolidate update modules to one module
- use logger instead of print for display functions
