--- src/my_unicorn/core/update/__init__.py
@@ -1,29 +1,0 @@
-"""Update management for installed AppImage applications."""
-
-from my_unicorn.core.update.display_update import (
-    display_check_results,
-    display_invalid_apps,
-    display_update_details,
-    display_update_error,
-    display_update_progress,
-    display_update_results,
-    display_update_success,
-    display_update_summary,
-    display_update_warning,
-)
-from my_unicorn.core.update.info import UpdateInfo
-from my_unicorn.core.update.manager import UpdateManager
-
-__all__ = [
-    "UpdateInfo",
-    "UpdateManager",
-    "display_check_results",
-    "display_invalid_apps",
-    "display_update_details",
-    "display_update_error",
-    "display_update_progress",
-    "display_update_results",
-    "display_update_success",
-    "display_update_summary",
-    "display_update_warning",
-]
