--- src/my_unicorn/core/update/catalog_cache.py
@@ -1,106 +1,0 @@
-"""In-memory catalog cache for update sessions.
-
-This module provides catalog caching functionality to reduce redundant file I/O
-when multiple apps share the same catalog.
-"""
-
-from __future__ import annotations
-
-import asyncio
-from typing import TYPE_CHECKING, Any
-
-from my_unicorn.constants import ERROR_CATALOG_MISSING
-from my_unicorn.exceptions import UpdateError
-from my_unicorn.logger import get_logger
-
-if TYPE_CHECKING:
-    from my_unicorn.config import ConfigManager
-
-logger = get_logger(__name__)
-
-
-class CatalogCache:
-    """In-memory catalog cache for update sessions.
-
-    Thread Safety:
-        - Safe for concurrent access across multiple asyncio tasks
-        - Protected by asyncio.Lock for concurrent access
-
-    Performance:
-        - First load: ~1-2ms (file I/O + JSON parse + validation)
-        - Cached load: ~0.01ms (dict lookup)
-        - Benefit: 100x faster for shared catalogs
-
-    """
-
-    def __init__(self, config_manager: ConfigManager) -> None:
-        """Initialize catalog cache.
-
-        Args:
-            config_manager: Configuration manager instance
-
-        """
-        self.config_manager = config_manager
-        self._cache: dict[str, dict[str, Any] | None] = {}
-        self._lock = asyncio.Lock()
-
-    async def load_catalog(self, ref: str) -> dict[str, Any] | None:
-        """Load catalog with in-memory caching.
-
-        This cache persists for the lifetime of the cache instance,
-        reducing redundant file I/O when multiple apps share the same catalog.
-        Uses asyncio.Lock to ensure thread-safe concurrent access.
-
-        Args:
-            ref: Catalog reference name (e.g., "qownnotes")
-
-        Returns:
-            Catalog entry dict or None if not found
-
-        Raises:
-            FileNotFoundError: If catalog file not found
-            ValueError: If catalog JSON is invalid or malformed
-
-        """
-        async with self._lock:
-            if ref not in self._cache:
-                entry = self.config_manager.load_catalog(ref)
-                # Cache result (even if None) to avoid repeated lookups
-                self._cache[ref] = entry  # type: ignore[assignment]
-
-            return self._cache.get(ref)
-
-    async def load_catalog_if_needed(
-        self, app_name: str, catalog_ref: str | None
-    ) -> dict[str, Any] | None:
-        """Load catalog entry if app references one.
-
-        Args:
-            app_name: Name of the app
-            catalog_ref: Catalog reference from app config
-
-        Returns:
-            Catalog entry dict or None
-
-        Raises:
-            UpdateError: If catalog reference is invalid or not found
-
-        """
-        if not catalog_ref:
-            return None
-
-        try:
-            return await self.load_catalog(catalog_ref)
-        except (FileNotFoundError, ValueError) as e:
-            msg = ERROR_CATALOG_MISSING.format(
-                app_name=app_name, catalog_ref=catalog_ref
-            )
-            raise UpdateError(
-                message=msg,
-                context={"app_name": app_name, "catalog_ref": catalog_ref},
-                cause=e,
-            ) from e
-
-    def clear(self) -> None:
-        """Clear the catalog cache."""
-        self._cache.clear()
